// src/main/webapp/src/app/rest/model/issueHistoryItemDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface IssueHistoryItemDTO { 
    field?: string;
    fieldtype?: string;
    from?: string;
    fromString?: string;
    to?: string;
    toString?: string;
}